Reorganise cluster tests
Aside from the upgrade tests any individual group here should only take a long time,
not a really painfully long time.

The naming structure should also make it easier to run, e.g. all compact clusters.
