+++ cosmo_tester/test_suites/cluster/compact_cfy_cluster_manager_upgrade_test.py
