+++ cosmo_tester/test_suites/cluster/full_cfy_cluster_manager_upgrade_test.py
